+++ server.js
@@ -1,6 +1,6 @@
 const app = require('./lib/app');
 const http = require('http');
-const port = process.env.PORT || 3000;
+const port = process.env.PORT || 8000;
 // const debug = require('debug')('myapp.server');
 
 const server = http.createServer(app);
